display refactor
